Applying use PHP namespaces properly for API
