+++ src/Controller/Api/ClinicController.php
@@ -2,10 +2,9 @@
 
 namespace App\Controller;
 
-use App\Entity\Veterinary;
 use App\Entity\Clinic;
 use App\Entity\Address;
-use App\Form\VeterinaryType;
+use App\Form\ClinicType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Routing\Annotation\Route;
@@ -16,35 +15,34 @@ use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 use Symfony\Component\Serializer\Serializer;
 use Symfony\Component\HttpFoundation\Response;
 
-class ApiVeterinaryController extends AbstractController
+class ClinicController extends AbstractController
 {
     public function list()
     {
-        $veterinaries = $this->getDoctrine()
-            ->getRepository(Veterinary::class)
+        $clinics = $this->getDoctrine()->getRepository(Clinic::class)
             ->findAll();
         
         $encoders = [new XmlEncoder(), new JsonEncoder()];
         $normalizers = [new ObjectNormalizer()];
        
         $serializer = new Serializer($normalizers, $encoders);
-        $jsonContent = $serializer->serialize($veterinaries, 'json', [
-            'ignored_attributes' => ['clinic']
+        $jsonContent = $serializer->serialize($clinics, 'json', [
+            'ignored_attributes' => ['veterinaries']
         ]);
 
         return new Response($jsonContent);
     }
 
-    public function show(Veterinary $veterinary)
+    public function show(Clinic $clinic)
     {
         $encoders = [new XmlEncoder(), new JsonEncoder()];
         $normalizers = [new ObjectNormalizer()];
        
         $serializer = new Serializer($normalizers, $encoders);
-        $jsonContent = $serializer->serialize($veterinary, 'json', [
-            'ignored_attributes' => ['clinic']
+        $jsonContent = $serializer->serialize($clinic, 'json', [
+            'ignored_attributes' => ['veterinaries']
         ]);
-
+        
         return new Response($jsonContent);
     }
 
@@ -55,60 +53,50 @@ class ApiVeterinaryController extends AbstractController
         $address->setNumber($request->get('number'));
         $address->setCity($request->get('city'));
         
-        $clinic = $this->getDoctrine()->getRepository('App\Entity\Clinic')
-            ->find($request->get('clinic'));
-
-        $veterinary = new Veterinary();
-        $veterinary->setName($request->get('name'));
-        $veterinary->setCrmv($request->get('crmv'));
-        $veterinary->setAddress($address);
-        $veterinary->addClinic($clinic);
+        $clinic = new Clinic();
+        $clinic->setName($request->get('name'));
+        $clinic->setAddress($address);
         
-        $form = $this->createForm(VeterinaryType::class, $veterinary);
-        $form->submit($veterinary);
+        $form = $this->createForm(ClinicType::class, $clinic);
+        $form->submit($clinic);
         
         $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->persist($veterinary);
+        $entityManager->persist($clinic);
         $entityManager->flush();
         
-        $response = new JsonResponse(['msg'=>'Veterinary created whit success!'], Response::HTTP_CREATED);
+        $response = new JsonResponse(['msg'=>'Clinic created whit success!'], Response::HTTP_CREATED);
         
         return $response;
     }
     
-    public function edit(Request $request, Veterinary $veterinary)
+    public function edit(Request $request, Clinic $clinic)
     {
-        $address = new Address();
+        $address = $clinic->getAddress();
         $address->setStreet($request->get('street'));
         $address->setNumber($request->get('number'));
         $address->setCity($request->get('city'));
         
-        $clinic = $this->getDoctrine()->getRepository('App\Entity\Clinic')
-            ->find($request->get('clinic'));
+        $clinic->setName($request->get('name'));
+        $clinic->setAddress($address);
 
-        $veterinary->setName($request->get('name'));
-        $veterinary->setCrmv($request->get('crmv'));
-        $veterinary->setAddress($address);
-        $veterinary->addClinic($clinic);
-        
-        $form = $this->createForm(VeterinaryType::class, $veterinary);
-        $form->submit($veterinary);
+        $form = $this->createForm(ClinicType::class, $clinic);
+        $form->submit($clinic);
         
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->flush();
         
-        $response = new JsonResponse(['msg'=>'Veterinary edited whit success!'], Response::HTTP_OK);
+        $response = new JsonResponse(['msg'=>'Clinic edited whit success!'], Response::HTTP_OK);
         
         return $response;
     }
 
-    public function delete(Veterinary $veterinary)
+    public function delete(Clinic $clinic)
     {
         $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($veterinary);
+        $entityManager->remove($clinic);
         $entityManager->flush();
 
-        $response = new JsonResponse(['msg'=>'Veterinary deleted whit success!'], Response::HTTP_OK);
+        $response = new JsonResponse(['msg'=>'Clinic deleted whit success!'], Response::HTTP_OK);
         
         return $response;
     }
