+++ src/Controller/Api/PetController.php
@@ -15,7 +15,7 @@ use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 use Symfony\Component\Serializer\Serializer;
 use Symfony\Component\HttpFoundation\Response;
 
-class ApiPetController extends AbstractController
+class PetController extends AbstractController
 {
     public function list()
     {
